add render folder and rearrange functions and files
