+++ src/render/render_utils.c
@@ -1,4 +1,4 @@
-#include "../includes/minirt.h"
+#include "../../includes/minirt.h"
 
 void	my_mlx_pixel_put(t_data *data, int x, int y, int color)
 {
