+++ trivec/cross_vecs.c
@@ -1,0 +1,9 @@
+#include "trivec.h"
+
+t_vec3d cross_vecs(t_vec3d *a, t_vec3d *b) {
+    t_vec3d result;
+    result.x = a->y * b->z - a->z * b->y;
+    result.y = a->z * b->x - a->x * b->z;
+    result.z = a->x * b->y - a->y * b->x;
+    return result;
+}
